# Conflicts:
#	myopia-business/management/src/main/java/com/wupol/myopia/business/management/domain/mapper/VisionScreeningResultMapper.java
#	myopia-business/management/src/main/java/com/wupol/myopia/business/management/domain/model/DistrictAttentiveObjectsStatistic.java
#	myopia-business/management/src/main/java/com/wupol/myopia/business/management/domain/model/DistrictMonitorStatistic.java
#	myopia-business/management/src/main/java/com/wupol/myopia/business/management/domain/model/SchoolVisionStatistic.java
#	myopia-business/management/src/main/resources/mapping/SchoolVisionStatisticMapper.xml
#	myopia-business/management/src/main/resources/mapping/VisionScreeningResultMapper.xml

+++ myopia-business/management/src/main/java/com/wupol/myopia/business/management/domain/model/DistrictAttentiveObjectsStatistic.java
@@ -5,13 +5,12 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.wupol.myopia.business.management.constant.WarningLevel;
-import jodd.util.MathUtil;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
 
 import java.io.Serializable;
-import java.util.Collections;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
@@ -63,7 +62,7 @@ public class DistrictAttentiveObjectsStatistic implements Serializable {
     /**
      * 重点视力对象--零级预警比例（均为整数，如10.01%，数据库则是1001）
      */
-    private Integer visionLabel0Ratio;
+    private BigDecimal visionLabel0Ratio;
     /**
      * 重点视力对象--一级预警人数（默认0）
      */
@@ -72,7 +71,7 @@ public class DistrictAttentiveObjectsStatistic implements Serializable {
     /**
      * 重点视力对象--一级预警比例（均为整数，如10.01%，数据库则是1001）
      */
-    private Integer visionLabel1Ratio;
+    private BigDecimal visionLabel1Ratio;
 
     /**
      * 重点视力对象--二级预警人数（默认0）
@@ -82,7 +81,7 @@ public class DistrictAttentiveObjectsStatistic implements Serializable {
     /**
      * 重点视力对象--二级预警比例（均为整数，如10.01%，数据库则是1001）
      */
-    private Integer visionLabel2Ratio;
+    private BigDecimal visionLabel2Ratio;
 
     /**
      * 重点视力对象--三级预警人数（默认0）
@@ -92,7 +91,7 @@ public class DistrictAttentiveObjectsStatistic implements Serializable {
     /**
      * 重点视力对象--三级预警比例（均为整数，如10.01%，数据库则是1001）
      */
-    private Integer visionLabel3Ratio;
+    private BigDecimal visionLabel3Ratio;
 
     /**
      * 重点视力对象--重点视力对象数量（默认0）
